requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
twython
pandas
numpy
nltk
matplotlib
textblob
wordcloud
sklearn